added frontend changes

+++ server/controllers/product.controller.js
@@ -50,11 +50,16 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
     const {id} = req.params;
+
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({success: false, message:"Product not found"});
+    }
+
     try{
         await Product.findByIdAndDelete(id);
         res.status(200).json({success: true, message:"Product deleted"});
     }catch (e){
         console.error('Error deleting Product: ' + e.message);
-        return res.status(404).json({success: false, message:'Product not found'});
+        return res.status(500).json({success: false, message:'Server Error: ' + e.message});
     }
 }
